# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API do Ageniza",
    "version": "0.1.0",
    "description": "API HTTP do Ageniza. A sessão é um cookie httpOnly emitido por `POST /auth/login`;\nas rotas autenticadas usam esse cookie e nunca um token no corpo ou na URL.\nErros seguem o mesmo corpo: `{ error: { code, message }, meta: { requestId } }`.\nA permissão exigida por cada rota aparece na extensão `x-permission` e no resumo em Markdown."
  },
...
  "components": {
    "schemas": {
...
      "createMediaUploadResponse": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          },
          "objectKey": {
            "type": "string",
            "minLength": 1
          },
          "category": {
            "type": "string",
            "enum": [
              "image",
              "video"
            ]
          },
          "upload": {
            "oneOf": [
              {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "single"
                    ]
                  },
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "type",
                  "url",
                  "expiresAt"
                ],
                "additionalProperties": false
              },
              {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "multipart"
                    ]
                  },
                  "uploadId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "partSizeBytes": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  },
                  "partCount": {
                    "type": "integer",
                    "exclusiveMinimum": 0
                  }
                },
                "required": [
                  "type",
                  "uploadId",
                  "partSizeBytes",
                  "partCount"
                ],
                "additionalProperties": false
              }
            ]
          }
        },
        "required": [
          "assetId",
          "objectKey",
          "category",
          "upload"
        ],
        "additionalProperties": false
      },
      "createMediaUploadRequest": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "contentType": {
            "type": "string",
            "minLength": 1,
            "maxLength": 256
          },
          "declaredSizeBytes": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 10737418240
          }
        },
        "required": [
          "fileName",
          "contentType",
          "declaredSizeBytes"
        ],
        "additionalProperties": false
      },
...
      "getMediaDownloadUrlResponse": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "format": "uri"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "url",
          "expiresAt"
        ],
        "additionalProperties": false
      }
    },
...
  },
  "paths": {
...
    "/agencies/{agencyId}/media/uploads": {
      "post": {
        "operationId": "createMediaUpload",
        "summary": "Inicia um upload de mídia",
        "description": "Devolve a URL assinada de PUT ou o plano de multipart; o tipo real é validado no fim, pelo conteúdo.\n\nAcesso: Sessão + vínculo com a agência. Permissão exigida: `midia.enviar`.",
        "tags": [
          "media"
        ],
        "x-permission": "midia.enviar",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "agencyId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createMediaUploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload iniciado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/createMediaUploadResponse"
                },
                "example": {
                  "assetId": "44444444-4444-4444-8444-444444444444",
                  "objectKey": "11111111-1111-4111-8111-111111111111/44444444-4444-4444-8444-444444444444/original.png",
                  "category": "image",
                  "upload": {
                    "type": "single",
                    "url": "https://storage.exemplo.test/arquivo.png?assinatura=ficticia",
                    "expiresAt": "2026-09-29T12:15:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Request validation failed"
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHENTICATED",
                    "message": "Authentication is required."
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "403": {
            "description": "CSRF_REJECTED: Request origin is not allowed | FORBIDDEN: You do not have permission to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "CSRF_REJECTED",
                    "message": "Request origin is not allowed"
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: Agency not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "Agency not found."
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "409": {
            "description": "QUOTA_EXCEEDED: This agency has reached its storage quota.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "QUOTA_EXCEEDED",
                    "message": "This agency has reached its storage quota."
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "413": {
            "description": "PAYLOAD_TOO_LARGE: Request payload is too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "PAYLOAD_TOO_LARGE",
                    "message": "Request payload is too large"
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "415": {
            "description": "UNSUPPORTED_MEDIA_TYPE: This content type is not accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNSUPPORTED_MEDIA_TYPE",
                    "message": "This content type is not accepted."
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR: An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INTERNAL_ERROR",
                    "message": "An unexpected error occurred"
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/agencies/{agencyId}/media/{assetId}/download-url": {
      "get": {
        "operationId": "getMediaDownloadUrl",
        "summary": "Emite uma URL assinada de leitura",
        "description": "Válida por poucos minutos, só GET, e apenas para um objeto confirmado da agência.\n\nAcesso: Sessão + vínculo com a agência. Permissão exigida: `midia.enviar`.",
        "tags": [
          "media"
        ],
        "x-permission": "midia.enviar",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "agencyId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "assetId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "original",
                "thumbnail",
                "preview"
              ],
              "default": "original"
            },
            "required": false,
            "name": "variant",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "URL assinada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/getMediaDownloadUrlResponse"
                },
                "example": {
                  "url": "https://storage.exemplo.test/arquivo.png?assinatura=ficticia",
                  "expiresAt": "2026-09-29T12:05:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VALIDATION_ERROR",
                    "message": "Request validation failed"
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: Authentication is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "UNAUTHENTICATED",
                    "message": "Authentication is required."
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "403": {
            "description": "FORBIDDEN: You do not have permission to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "FORBIDDEN",
                    "message": "You do not have permission to perform this action."
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: Media asset not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "NOT_FOUND",
                    "message": "Media asset not found."
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "409": {
            "description": "VARIANT_NOT_READY: This variant has not been generated yet. | VARIANT_PROCESSING_FAILED: Video processing failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "VARIANT_NOT_READY",
                    "message": "This variant has not been generated yet."
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR: An unexpected error occurred",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "INTERNAL_ERROR",
                    "message": "An unexpected error occurred"
                  },
                  "meta": {
                    "requestId": "req-de-exemplo"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}